Sync 9.1

+++ test/aspnet-core-demo/AbpAspNetCoreDemo/Controllers/ProductsDtoController.cs
@@ -7,9 +7,9 @@ using AbpAspNetCoreDemo.Core.Domain;
 
 namespace AbpAspNetCoreDemo.Controllers
 {
-    public class ProductsDtoController : AbpODataDtoController<Product, ProductDto, ProductCreateInput, long>, ITransientDependency
+    public class ProductsDtoController : AbpODataDtoController<Product, ProductDto, ProductCreateInput, int>, ITransientDependency
     {
-        public ProductsDtoController(IRepository<Product, long> repository, IObjectMapper objectMapper) : base(repository, objectMapper)
+        public ProductsDtoController(IRepository<Product, int> repository, IObjectMapper objectMapper) : base(repository, objectMapper)
         {
             GetPermissionName = "GetProductPermission";
             GetAllPermissionName = "GetAllProductsPermission";
